pslbackend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
base58==2.1.1
bcrypt==4.3.0
bitarray==3.7.1
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
ckzg==2.1.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.7
cytoolz==1.0.1
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
faiss-cpu==1.12.0
fastapi==0.116.1
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.4
groq==0.31.1
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.1
idna==3.10
ImageHash==4.3.2
iniconfig==2.1.0
ipfshttpclient==0.7.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-groq==0.3.8
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.11
langsmith==0.4.30
mangum==0.19.0
MarkupSafe==3.0.2
marshmallow==3.26.1
motor==3.7.1
mpmath==1.3.0
multiaddr==0.0.9
multidict==6.6.4
mypy_extensions==1.1.0
netaddr==1.3.0
networkx==3.5
nltk==3.9.2
numpy==2.3.3
openai==1.107.1
orjson==3.11.3
packaging==25.0
parsimonious==0.10.0
passlib==1.7.4
pillow==11.3.0
pluggy==1.6.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pymongo==4.14.1
pyparsing==3.2.3
pypdf==6.1.0
pytest==8.4.2
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pyunormalize==16.0.0
PyWavelets==1.9.0
pywin32==311
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
rlp==4.1.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.1
sentence-transformers==5.1.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
toolz==1.0.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.56.2
types-requests==2.32.4.20250809
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
varint==1.0.2
web3==7.13.0
websockets==15.0.1
yarl==1.20.1
zstandard==0.25.0
qrcode==8.2
pyotp==2.9.0
instagrapi==2.0.0
py-algorand-sdk==2.4.0

